+++ libs/Symfony/Component/Process/Pipes/WindowsPipes.php
@@ -67,16 +67,16 @@ class WindowsPipes extends AbstractPipes
             set_error_handler(function ($type, $msg) use (&$lastError) {
                 $lastError = $msg;
             });
-            for ($i = 0; ; $i++) {
+            for ($i = 0;; $i++) {
                 foreach ($pipes as $pipe => $name) {
                     $file = sprintf('%s\\sf_proc_%02X.%s', $tmpDir, $i, $name);
 
-                    if (! $h = fopen($file.'.lock', 'w')) {
-                        if (file_exists($file.'.lock')) {
+                    if (! $h = fopen($file . '.lock', 'w')) {
+                        if (file_exists($file . '.lock')) {
                             continue 2;
                         }
                         restore_error_handler();
-                        throw new RuntimeException('A temporary file could not be opened to write the process output: '.$lastError);
+                        throw new RuntimeException('A temporary file could not be opened to write the process output: ' . $lastError);
                     }
                     if (! flock($h, LOCK_EX | LOCK_NB)) {
                         continue 2;
@@ -107,12 +107,12 @@ class WindowsPipes extends AbstractPipes
 
     public function __sleep(): array
     {
-        throw new BadMethodCallException('Cannot serialize '.__CLASS__);
+        throw new BadMethodCallException('Cannot serialize ' . __CLASS__);
     }
 
     public function __wakeup()
     {
-        throw new BadMethodCallException('Cannot unserialize '.__CLASS__);
+        throw new BadMethodCallException('Cannot unserialize ' . __CLASS__);
     }
 
     public function __destruct()
@@ -157,7 +157,7 @@ class WindowsPipes extends AbstractPipes
             if ($w) {
                 @stream_select($r, $w, $e, 0, Process::TIMEOUT_PRECISION * 1E6);
             } elseif ($this->fileHandles) {
-                usleep(Process::TIMEOUT_PRECISION * 1E6);
+                usleep((int) (Process::TIMEOUT_PRECISION * 1E6));
             }
         }
         foreach ($this->fileHandles as $type => $fileHandle) {
